Miliseconds extension moved to file
