+++ AirCasting/SessionsSynchronization/Controller/Interfaces/SessionUpstream.swift
@@ -58,15 +58,3 @@ extension SessionsSynchronization {
         let time: Date
     }
 }
-
-extension Date {
-    static let msFormatter: DateFormatter = {
-        let f = DateFormatter()
-        f.dateFormat = "SSS"
-        return f
-    }()
-    
-    var miliseconds: Int {
-        Int(Date.msFormatter.string(from: self).dropFirst())!
-    }
-}
